fix: change duration field type to tinyint in review_logs table

+++ docker/mysql/mysql-init-scripts/000_schema.sql
@@ -178,7 +178,7 @@ CREATE TABLE IF NOT EXISTS review_logs (
 
     learning_steps INT NOT NULL,
     state TINYINT NOT NULL,
-    duration INT NOT NULL,
+    duration TINYINT NOT NULL,
 
     PRIMARY KEY (user_id, id),
     INDEX ix_review_logs_usn (user_id ASC, usn ASC)
